pyproject.toml
[project]
name = "inktvis"
version = "0.1.0"
description = "Convert scanned book pages to structured Markdown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
]

[project.scripts]
inktvis = "inktvis.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
